Refactor: export message component
