+++ src/LoginPage.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Box, Typography } from '@mui/material';
+
+const LoginPage: React.FC = () => (
+  <Box p={2}>
+    <Typography variant="h4">Login</Typography>
+    <Typography>Login page will appear here.</Typography>
+  </Box>
+);
+
+export default LoginPage;
