+++ src/components/Contact.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Box, Typography, Card, CardActionArea, CardContent, Avatar, Stack } from '@mui/material';
-import { Contact } from '../types';
+import { Contact } from './types';
 
 const ContactCard: React.FC<{ contact: Contact, handleClick: (id: string) => void }> = ({ contact, handleClick }) => {
     return <Card
