+++ src/components/Stories.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Box, Avatar, Typography, IconButton } from '@mui/material';
 import AddIcon from '@mui/icons-material/Add';
-import { Contact } from '../types';
+import { Contact } from './types';
 import StoryViewer from './StoryViewer';
 
 interface StoriesProps {
